refactor(ui): extract the diff line computation into a pure helper
